consolidate the QTC and QSO logbook structs
