+++ core/entry/entry_test.go
@@ -275,7 +275,7 @@ func TestEntryController_LogNewQSO(t *testing.T) {
 	log.Activate()
 	log.On("NextNumber").Return(core.QSONumber(1))
 	log.On("LastExchange").Return([]string{"599", "001", ""})
-	log.On("Log", qso).Once()
+	log.On("LogQSO", qso).Once()
 	qsoList.Activate()
 	qsoList.On("FindDuplicateQSOs", dl1abc, mock.Anything, mock.Anything).Return([]core.QSO{})
 	qsoList.On("SelectLastQSO").Twice()
@@ -488,7 +488,7 @@ func TestEntryController_LogDuplicateQSO(t *testing.T) {
 	log.Activate()
 	log.On("NextNumber").Return(core.QSONumber(2))
 	log.On("LastExchange").Return([]string{"599", "001", ""})
-	log.On("Log", dupe).Once()
+	log.On("LogQSO", dupe).Once()
 	qsoList.Activate()
 	qsoList.On("FindDuplicateQSOs", dl1abc, mock.Anything, mock.Anything).Return([]core.QSO{qso})
 	qsoList.On("SelectLastQSO").Twice()
@@ -585,7 +585,7 @@ func TestEntryController_EditQSO(t *testing.T) {
 	controller.Enter("B02")
 
 	log.Activate()
-	log.On("Log", changedQSO).Once()
+	log.On("LogQSO", changedQSO).Once()
 	log.On("LastExchange").Return([]string{"599", "001", ""}).Times(2)
 	log.On("NextNumber").Return(core.QSONumber(35))
 	controller.EnterPressed()
