+++ core/logbook/dupe_test.go
@@ -1,4 +1,4 @@
-package qso
+package logbook
 
 import (
 	"testing"
