+++ core/logbook/logbook_test.go
@@ -1,13 +1,13 @@
-package qso
+package logbook
 
 import (
 	"testing"
 	"time"
 
+	"github.com/ftl/hamradio/callsign"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/ftl/hamradio/callsign"
 	"github.com/ftl/hellocontest/core"
 	"github.com/ftl/hellocontest/core/clock"
 	"github.com/ftl/hellocontest/core/mocked"
@@ -17,17 +17,25 @@ func TestNew(t *testing.T) {
 	logbook := New(clock.New())
 
 	assert.Equal(t, core.QSONumber(1), logbook.NextNumber(), "next number of empty log should be 1")
-	assert.Empty(t, logbook.All(), "empty log should not contain any QSO")
+	assert.Empty(t, logbook.AllQSOs(), "empty log should not contain any QSO")
 }
 
 func TestLoad(t *testing.T) {
 	qsos := []core.QSO{
 		{MyNumber: 123},
 	}
+	now := time.Now()
+	qtcs := []core.QTC{
+		{QSONumber: 234, Timestamp: now.Add(-1 * time.Minute)},
+		{QSONumber: 235, Timestamp: now},
+	}
 
-	logbook := LoadLogbook(clock.New(), qsos)
+	logbook := Load(clock.New(), qsos, qtcs)
 
 	assert.Equal(t, core.QSONumber(124), logbook.NextNumber())
+
+	actualQTCs := logbook.AllQTCs()
+	assert.Equal(t, qtcs, actualQTCs)
 }
 
 func TestLog_Log(t *testing.T) {
@@ -36,10 +44,10 @@ func TestLog_Log(t *testing.T) {
 	logbook := New(clock)
 
 	qso := core.QSO{MyNumber: 1}
-	logbook.Log(qso)
+	logbook.LogQSO(qso)
 
-	require.Equal(t, 1, len(logbook.All()), "after logging one QSO, the log should have one item")
-	loggedQso := logbook.All()[0]
+	require.Equal(t, 1, len(logbook.AllQSOs()), "after logging one QSO, the log should have one item")
+	loggedQso := logbook.AllQSOs()[0]
 	assert.Equal(t, now, loggedQso.LogTimestamp, "LogTimestamp is wrong")
 }
 
@@ -51,14 +59,14 @@ func TestLog_LogAgain(t *testing.T) {
 
 	clock.On("Now").Once().Return(now)
 	qso := core.QSO{MyNumber: 1, TheirNumber: 1}
-	logbook.Log(qso)
+	logbook.LogQSO(qso)
 
 	clock.On("Now").Once().Return(then)
 	qso.TheirNumber = 2
-	logbook.Log(qso)
+	logbook.LogQSO(qso)
 
-	require.Equal(t, 2, len(logbook.All()), "log should have two items")
-	lastQso := logbook.All()[1]
+	require.Equal(t, 2, len(logbook.AllQSOs()), "log should have two items")
+	lastQso := logbook.AllQSOs()[1]
 	assert.Equal(t, then, lastQso.LogTimestamp, "last item should have last timestamp")
 	assert.Equal(t, core.QSONumber(2), lastQso.TheirNumber, "last item should have latest data")
 }
@@ -73,7 +81,7 @@ func TestLog_EmitRowAdded(t *testing.T) {
 	}))
 
 	qso := core.QSO{MyNumber: 1}
-	logbook.Log(qso)
+	logbook.LogQSO(qso)
 
 	assert.True(t, emitted)
 }
@@ -82,7 +90,7 @@ func TestLog_NextNumber(t *testing.T) {
 	logbook := New(clock.New())
 
 	qso := core.QSO{MyNumber: 123}
-	logbook.Log(qso)
+	logbook.LogQSO(qso)
 
 	assert.Equal(t, core.QSONumber(124), logbook.NextNumber(), "next number should be the highest existing number + 1")
 }
@@ -91,7 +99,7 @@ func TestLastBand(t *testing.T) {
 	logbook := New(clock.New())
 	assert.Equal(t, core.NoBand, logbook.LastBand())
 
-	logbook.Log(core.QSO{Callsign: callsign.MustParse("DL1ABC"), MyNumber: 1, Band: core.Band80m})
+	logbook.LogQSO(core.QSO{Callsign: callsign.MustParse("DL1ABC"), MyNumber: 1, Band: core.Band80m})
 	assert.Equal(t, core.Band80m, logbook.LastBand())
 }
 
@@ -99,6 +107,6 @@ func TestLastMode(t *testing.T) {
 	logbook := New(clock.New())
 	assert.Equal(t, core.NoMode, logbook.LastMode())
 
-	logbook.Log(core.QSO{Callsign: callsign.MustParse("DL1ABC"), MyNumber: 1, Mode: core.ModeDigital})
+	logbook.LogQSO(core.QSO{Callsign: callsign.MustParse("DL1ABC"), MyNumber: 1, Mode: core.ModeDigital})
 	assert.Equal(t, core.ModeDigital, logbook.LastMode())
 }
